Push

+++ Contagion/Contagion.Storage/Models/User.cs
@@ -2,7 +2,7 @@ namespace Contagion.Storage.Models
 {
   public class User
   {
-    public int UserPhone { get; set; }
+    public long UserPhone { get; set; }
     public decimal Lat { get; set; }
     public decimal Long { get; set; }
   }
